Changed verything to int mintues

+++ acmeflix-app/src/main/java/com/acmeflix/transfer/resource/ProfileResource.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 public class ProfileResource extends BaseResource {
     private Long profileID;
     private String profileName;
-    private Double viewedMinutes;
+    private Integer viewedMinutes;
 }
